Moved files to css folder
