+++ client/src/components/IconButton.js
@@ -11,7 +11,7 @@
 
 import React from "react";
 import background from "../images/greenbutton.png";
-import "../styles/iconbutton.css";
+import "../css/iconbutton.css";
 
 function IconButton(props) {
   return (
