+++ client/src/css/header.css
@@ -5,6 +5,7 @@
   grid-template-columns: 1.5fr 4fr 1fr;
   column-gap: 5%;
   height: 130px;
+  width: 100%;
   box-sizing: border-box;
 }
 
@@ -51,6 +52,7 @@
   grid-column: unset;
   display: none;
   padding-top: 63px;
+  padding-right: 0px;
 }
 
 @media only screen and (max-width: 750px) {
@@ -66,7 +68,6 @@
     display: block;
     grid-column: 2;
     width: 43px;
-    padding-right: 15px;
   }
   .header-welcome-text {
     font-size: 20px;
@@ -80,11 +81,10 @@
   }
   #header {
     height: 97px;
-    grid-template-columns: auto 43px;
+    grid-template-columns: auto 55px;
   }
   .header-welcome img {
-    padding-top: 47px;
     width: 47px;
-    height: 40px;
+    height: inherit;
   }
 }
